fix: プレビューの縦横比を維持するように
