+++ src/renderer/App.tsx
@@ -1,14 +1,11 @@
-import FrameBufferRenderer from "./frameRenderer/FrameBufferRenderer";
-import FrameTextureRenderer from "./frameRenderer/FrameTextureRenderer";
 import Dock from "./dock/Dock";
+import FrameRenderer from "./frameRenderer/FrameRenderer";
 import Timeline from "./timeline/Timeline";
-
-const config = await window.main.getConfig();
 
 function App() {
   return (
     <Dock>
-      {config.fastPreview ? <FrameTextureRenderer /> : <FrameBufferRenderer />}
+      <FrameRenderer />
       <Timeline />
     </Dock>
   );
