+++ src/renderer/frameRenderer/useFrameBufferRenderer.ts
@@ -21,7 +21,7 @@ fn main(@location(0) texCoord: vec2f) -> @location(0) vec4f {
 const FRAME_WIDTH = 1920;
 const FRAME_HEIGHT = 1080;
 
-const FrameBufferRenderer = () => {
+const useFrameBufferRenderer = () => {
   const { resources, canvas: canvasRef } = useBufferPreviewWebGPU({
     fragmentShaderCode,
     width: FRAME_WIDTH,
@@ -147,14 +147,7 @@ const FrameBufferRenderer = () => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [viewerState, resources]);
 
-  return (
-    <canvas
-      ref={canvasRef}
-      width={FRAME_WIDTH}
-      height={FRAME_HEIGHT}
-      style={{ width: "100%", height: "100%" }}
-    />
-  );
+  return canvasRef;
 };
 
-export default FrameBufferRenderer;
+export default useFrameBufferRenderer;
