+++ src/renderer/index.css
@@ -8,7 +8,6 @@
   min-height: 50px;
   display: flex;
   flex-direction: column;
-  padding: 8px;
   border: 1px solid #c0c0c0;
   border-top: none;
   background: white;
